move agreement utils to agreement file
