+++ packages/models/src/agreement/agreement.ts
@@ -154,3 +154,17 @@ export const AgreementInvolvedAttributes = z.object({
 export type AgreementInvolvedAttributes = z.infer<
   typeof AgreementInvolvedAttributes
 >;
+
+export const agreementUpdatableStates: AgreementState[] = [
+  agreementState.draft,
+];
+
+export const agreementUpgradableStates: AgreementState[] = [
+  agreementState.active,
+  agreementState.suspended,
+];
+
+export const agreementDeletableStates: AgreementState[] = [
+  agreementState.draft,
+  agreementState.missingCertifiedAttributes,
+];
