+++ packages/models/src/index.ts
@@ -1,5 +1,4 @@
 export * from "./agreement/agreement.js";
-export * from "./agreement/agreementUtils.js";
 export * from "./agreement/agreementEvents.js";
 export * from "./attribute/attribute.js";
 export * from "./errors.js";
